feat: Add palettes.json file with predefined color palettes
This commit adds a new file, palettes.json, to the setup directory. The file contains predefined color palettes for different themes, including "berry", "malory", and "neodutch". Each palette consists of an array of color codes. This addition will provide a convenient way to select and use predefined color schemes in the application.