feat: Add installs.fish function for installing software
This commit adds a new file, installs.fish, to the functions directory. The function allows for the installation of multiple pieces of software from any source while adding them to the path and keeping everything up to date. It includes options for installing with Snap, Homebrew, or the default package manager. The function also includes extra steps for specific programs like creating directories, adding to the PATH, and creating aliases. This addition provides a convenient way to install and manage software packages in the application.