feat: Add cpfunc to install.fish script
This commit adds the `cpfunc` function to the `install.fish` script. The `cpfunc` function is used to copy functions to the appropriate directory and make them executable. This addition ensures that the `cpfunc` function is available for use in the installation process.

Note: This commit message does not include any meta information like issue references, tags, or author names.