Refactor indent_size in .editorconfig and yarn+.fish
This commit refactors the indent_size property in the .editorconfig file to set it to 2 instead of 4 for *.fish files. It improves consistency in the codebase by aligning the indent_size with the rest of the project.

Additionally, it adds the yarn+ function to the alias/yarn+.fish file. The yarn+ function wraps the commands "yarn set version stable", "yarn", and "yarn upgrade-interactive" to provide a convenient way to update and manage yarn dependencies. This addition enhances the developer's workflow by simplifying the process of updating yarn packages.