Refactor n+ function to use asdf or nvm based on user preference
This commit modifies the n+ function in the alias/n+.fish file to use either asdf or nvm based on the user's preference. The function now checks the value of the "cauldron_packman_pref" variable and executes the corresponding update command for either asdf or nvm. This change enhances the flexibility of the n+ function by allowing users to choose their preferred package manager for updating Node and Yarn to the latest version.