feat: Add nvm_update_node function for managing Node.js versions
This commit adds the nvm_update_node function to the nvm_update_node.fish script. The function checks for the presence of a .nvmrc file or a ~/.nvmrc file to determine the desired Node.js version. If the version is not installed, it installs it using nvm. The function then sets the desired version as the default and current version of Node.js. This addition enhances the script's functionality by providing a convenient way to manage Node.js versions using nvm.