feat: Add getLatestGithubReleaseAsJSON function to retrieve the latest release of a GitHub repository
This commit adds the getLatestGithubReleaseAsJSON function to the repository. The function allows users to retrieve the latest release of a GitHub repository as JSON. It accepts a target owner / repository as an argument and makes a request to the GitHub API to fetch the latest release information. This will make creation of automatic updating tools much easier, as we can prompt the user with information from the latest release, when asking if they want to update.