feat: Add asdf_update_go and asdf_update_ruby scripts
The asdf_update_go script updates the Go version to the latest version using the asdf package manager. It also sets the updated version as the global Go version and checks if there are any local .rb files to set the updated version as the local Go version.

The asdf_update_ruby script updates the Ruby version to the latest version using the asdf package manager. It also sets the updated version as the global Ruby version and checks if there are any local .rb files or a .ruby-version file to set the updated version as the local Ruby version.